Move renderer to renderers.py
